Update CmdStanMLE.create_inits() compatibility
This change simplifies the create_inits() for the MLE
params to always return a dictionary of inits, rather
than a list for multiple chains. The default behavior
of the sample method is to initialize all chains
at the same init if only one is given per param.

The chains paramter is kept and the seed parameter is
added to the signature, despite being no-ops, for
the purposes of maintaining uniformity across the
other create_inits() methods on other stanfit objects.