Add StanCsvMCMC dataclass
This class is intended to serve as an abstraction layer between
the Stan fit objects (like CmdStanMCMC) and the Stan CSV output.
It converts the Stan CSV into structured format that can be used
to straightforwardly extract relevant info.

The practically important contribution here is the implementation
of a faster samples parser using pandas/pyarrow intended to replace
the pure-Python implementation.

Some of the structure/utilties included in this commit are intended
to clean up logic elsewhere in the library by sharing functionality
and making the processing of Stan CSV files more consistent across
the varying inference methods.