Make pathfinder default num_paths explicit
Previous implementation left num_paths as None by default.
This meant it would be set as a default value of 4 by cmdstan
under the hood. While functional, this meant that cmdstanpy
did not explicitly know the number of paths used when running
pathfinder. This made it awkward to write code that needed to
know how many paths were run.

By making the default explicitly 4 in cmdstanpy, this clarifies
behavior at the cost of not automaticallyu inheriting any changes
to default cmdstan behavior.