Lazily load metric info from file
In shifting the metric/type/step size info to being sourced
from the metric JSON files instead of the CSV comments, we no longer
will populate those attributes during the _assemble_draws step and
instead lazily load the attributes from the metric JSON files when
the properties are accessed.

This allows the draws to still be used/manipulated in cases when
the metric information is unavailable such as when only the Stan
CSV files are saved.